Moved out tests for ResNet
